# perf: unsafe.WriteSlice (#501)